Propagating errors is not an error.
Log spew when something goes wrong makes it hard to see which part is actually broken.

Reducing the severity of logging when we just propagate an error message along the chain of reporting. Errors that are user generated are not a sign that the monarch components that move the errors around are working sub-optiminally so anything over an `info` at this level is probably not right.

Removed at least one redundant log where we get two messages.

Ultimately, we should be checking errors in the global context on client script exit, and logging to the user console until that point.

Differential Revision: [D76377545](https://our.internmc.facebook.com/intern/diff/D76377545/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D76377545/)!

ghstack-source-id: 289532985
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/219